aso-service/requirements.txt
requests>=2.31.0
tqdm>=4.66.0
fastapi==0.111.0
uvicorn==0.29.0
pymysql==1.1.0
python-dotenv==1.0.0
